refactor: Client API 로직 통일
- Client에서 사용하는 API들이 같은 instance를 지니도록 통일
- Client API 관련 폴더 구조 개선
