+++ FE/src/pages/LoginPage/LoginPage.tsx
@@ -1,7 +1,7 @@
 import { useState, useContext } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useForm, SubmitHandler } from 'react-hook-form';
-import { postLogin } from '../../api/auth';
+import { postLogin } from '../../api';
 import { LoginFetchData as FormData } from 'src/types/type';
 import { Container, Inner, Form } from './LoginPage.styles';
 import { AuthContext } from '../../contexts/AuthContexts';
